move capa/features/extractors/__init__.py logic to base_extractor.py
