+++ capa/features/extractors/smda/extractor.py
@@ -4,7 +4,7 @@ import capa.features.extractors.smda.file
 import capa.features.extractors.smda.insn
 import capa.features.extractors.smda.function
 import capa.features.extractors.smda.basicblock
-from capa.features.extractors import FeatureExtractor
+from capa.features.extractors.base_extractor import FeatureExtractor
 
 
 class SmdaFeatureExtractor(FeatureExtractor):
